Optimize action section for mobile, for real
Action section shrinks much better, including its horizontal margins,
so that it looks reasonable even on a small device, and even though
there is now more text in the blue box. The button now expands flexibly
in both directions and fills in the horizontal space post-being-
wrapped-below-the-text.

Text is updated to give credit to Josh Comeau's Interactive Guide to
Flexbox, which helped me finally figure out how to do this properly.